feat: improve code format and add missing return
